refactor: move local.conf config into container
Added a README that describes the current process.

Reconfigure local.conf for ODL/non-ODL use inside the container so we
don't need to rebuild the image with each local.conf.  This follows the
general philosohpy that most configuration should be outside the
container image itself and shoul dbe done at runtime.

Added scripts to start the consul server and extract its IP. This info
needs to be fed back to the other docker daemons in the system as both
the --cluster-store and --cluster-advertise arguments to the daemon.