fix: enable devstack in service node without error
Major modification to local.conf: service.local.conf should now include
all of the configuration options that were present in the original
image. Some of the current options may be over-constrained and should be
relaxed, but not until setup is ironed out.
ISSUE: when neutron is stacking, it tries to create iptables rules for
Linux bridge if used as an ML2 mechanism_driver. I added the config
option "Q_USE_SECGROUP=False" which prevents interacting with the Linux
bridge.  This might have also been "worked around" when the ML2
mechanism is set to OVS only.

Move branch checkout into start.sh since it's closer to a runtime
configuration than something inherent to the image.

ISSUE: we hit a regression with setuptools 30.0 which apparently need
packaging===16.8, but it's locked at 16.7 in
requirements/upper-requirements.txt.  This prevents stacking and needs
to be fixed after the correct branch has been selected.
No fix is currently in place for this.