enable functional interactive instance console
Instance console was throwing errors complaining about 'unknown
keycodes'.  This was, apparently, caused by a change in noVNC that sent
additional keycode data. [1] [2]
The solution was to disable the vnc "keymap" in qemu.
Additional code in this commit enables waiting on a valid response from
Horizon or waiting for valid network response from the service node
before beginning stacking.
Finally, created a script-let that shows the proxied address of the
noVNC console after container port maps.